@@ -2,7 +2,7 @@ package resolvers
 
 // This file will be automatically regenerated based on the schema, any resolver implementations
 // will be copied through when generating and any unknown code will be moved to the end.
-// Code generated by github.com/99designs/gqlgen version v0.17.57
+// Code generated by github.com/99designs/gqlgen version v0.17.61
 
 import (
 	api "github.com/photoview/photoview/api/graphql"
